PID code

+++ src/CC1P1PiAnalysis.cpp
@@ -671,7 +671,7 @@ bool CC1P1PiAnalysis::FindParticles(Minerva::PhysicsEvent* event) const
                 
             }
             else if(pr_score_N < pi_score_N){
-                for(int hyp_counter = 0; hyp_counter < partHypVec.size(); hyp_counter++){
+                for(int hyp_counter = 0; hyp_counter < (int)partHypVec.size(); hyp_counter++){
                     if( partHypVec[hyp_counter]->idcode() == Minerva::Particle::Pion){
                         found_p_no = hyp_counter;
                     }
